Change function VariablesNoIteration
The code has an error when calling function "solveIteration", the issue was present due to the output of VariablesNoIteration that are numbers with a dot at the end (float numbers). Change on the function to create np.array of integers instead.
Root cause: failure in for i in VariablesNoIteration  due to the float numbers